requirements.txt
streamlit==1.31.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pyyaml==6.0.1
pandas==2.2.0
zhipuai==2.0.1
python-dotenv==1.0.1
